requirements.txt
# Core ML (nécessite GPU recommandé)
torch>=2.0.0
faster-whisper>=1.0.0
pyannote.audio>=3.1.0
silero-vad>=5.0.0
webrtcvad>=2.0.10
accelerate
bitsandbytes
lightning>=2.4
transformers>=4.45.0
mistral-common[audio]>=1.9.0
vosk
resemblyzer
diarize
speechbrain
onnxruntime

# Always required (CPU ok)
librosa>=0.10.0
markdown
soundfile>=0.12.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
keyring
python-multipart
scipy
pandas
numba
soxr
boto3
matplotlib
scikit-learn
jinja2

# Testing & Quality (Cadrage LLM)
pytest-cov
hypothesis
mutmut
pytest-snapshot
beautifulsoup4